Added scripts/lumi.py and associated files
This script calculates the integrated luminosity analyzed in CRAB tasks.
Run this script e.g. like this:
lumi.py crab_0_*
If test/lumi/lumi-map.txt is in a reasonable shape, this will then determine and print the integrated luminosity per task.
If you do not want to filter on the files/pattern in lumi-map.txt, use the -a option to simply calculate the total analyzed luminosity without any good run list.