music_crab: New way to pass the CMSSW config
music_crab now only takes one argument, which is the sample files.
The CMSSW config can either be given on the command line with the -c options,
or configured in the sample file.
For the latter, add at the beginning of the sample file a line like this:
config=PATH
If PATH is absolute, the file is simply takes as it is.
If path is relative, the file will be looked for relative to the config-file-path.
This is usually $CMSSW_BASE/src/MUSiCProject/Skimming/test/configs,
but can be changes using --config-dir option.