MUSiCSkimmer: Changed the way filters are handled.
In order to be more flexible, the outcome of all filters is stored as bools in
each event rather than throwing away the event.

Also changed the structure of the config files to improve clarity.
The functions that all config files have in common are now called in the
prepare() function of MUSiCSkimmer_cfi.py.
Thus changes have to be done only in one place and not in each config file
separately.