WIP: MUSiCSkimmer: Updated taus to 2012 recommendations.
Due to problems with the reconstruction of the tau energy the PF2PAT taus
cannot be used with CMSSW 5_3_X. For the latest stable recommendation see:
   https://twiki.cern.ch/twiki/bin/view/CMSPublic/SWGuidePFTauID?rev=145#5_3_12_and_higher
It still contains known problems for high pT taus. A newer version with fixes is
available from the Tau POG, but remains untested and not recommended.

Also added variables for cross-checking tau decay modes.