parseSampleList: Changed handling of timestamps.
Now if a timestamp is found in the file you are parsing, the samples
corresponding to this timestamp are grouped together. For each such group of
samples a single output file is written. This means you can run the script
several times, in case you are counting on new samples becoming complete over
time.
If you do not want to create a file for each group of timestamp sorted samples,
use the '--no-timestamp' option.