WIP: Tools: Removed checking of GT version and added newest GTs.
Since the global tag versions are not significant (n does not have to be less
favorable than n+1) and there are also subversions, e.g., v7A, it does not make
sense to compare the GT based on these version numbers.
Futhermore the newest GT are added for both Summer12 MC (CMSSW 5_3x) and 2012
prompt data. Notice that for the data the GT actually used during the skimming
are set in music_crab.py and are required by the ECal laser correction filter
individually for each run. The data GT in Tools.py is only a place holder.