WIP: MUSiCSkimmer: 2012 version (CMSSW 5_3_X) of rho for eff. area corrections.
For analysis of the 2012 dataset with CMSSW_5_3_X a new version of the rho
variable is recommended. See this Twiki and the corresponding config for
details:

https://twiki.cern.ch/twiki/bin/view/CMS/EgammaEARhoCorrection?rev=12#Rho_for_2012_Effective_Areas
http://cmssw.cvs.cern.ch/cgi-bin/cmssw.cgi/CMSSW/RecoJets/JetProducers/python/PFJetParameters_cfi.py?revision=1.10&view=markup

Three diferent versions of rho are now written into the event. The PF electron
isolation effective area correction now uses the 2012 version of rho (with
eta_max = 4.4).