WIP: MUSiCSkimmer: Check if the datastream contains any unprescaled triggers.
To insure that datastreams being used for our analysis (in the classification)
contain at least one unprescaled trigger a mapping of the triggers to their
datastreams is necessary. This lead to rather extensive changes in the trigger
code. In particular the filters can no longer be initialized in the same
function as the triggers.

The configuration file can now be used in three different ways to get the
triggers:

a) Specify explicit HLT paths
   Only these exact triggers and their corresponding datastreams will be
   considered, if they are unprescaled.
   You cannot additionally specify datastreams in this case.

b) Specify explicit datastreams
   All unprescaled triggers from these datastreams will be taken into
   consideration. If the datastream only contains prescaled triggers, its
   UserRecord will be set to false.

c) Specify nothing
   All datastreams and corresponding unprescaled triggers will be considered. If
   the datastream only contains prescaled triggers, its UserRecord will be set
   to false.