WIP: Tools: New version of MET corrections.
A new recipe for MET corrections was released for the git based CMSSW releases:
   https://twiki.cern.ch/twiki/bin/view/CMSPublic/WorkBookMetAnalysis?rev=170#How_to_apply_the_MET_corrections
This no longer runs via the 'usePF2PAT' function and is based on 'reco::MET'.
All different combinations of MET corrections are now saved for each event, to
ensure maximal flexibility.
The recommended MET corrections to use in your analysis are:
   Type-0 + Type-I + xy-Shift
The Type-0 correction has been implemented in two versions, which use different
methods to calculate charged pile-up particles. For MUSiC we want to use:
   'pfMetT0pcT1Txy'
which uses particle-flow charged hadrons for the Type-0 correction.