WIP: Added additional data streams to skimming configs.
The 'DoubleMuParked' stream of Run A has been found to be missing some (few)
events. It is recommended to use the 'DoubleMu' stream for Run A. For later
runs only the 'DoubleMuParked' stream is available.

Some datastreams are still commented-out to save time and resources during
skimming. Once there is more time comment them in and skim away!