First commit with changes to run the skimmer on 13 TeV samples.
The skimmer runs but some results are preliminary: No JEC are apllied
and the genJets are switched off. Also some filters are not running.
This should be updated once the new recipes are available. Also the
object reco and id variables have to be updated in the future.

Nevertheless, this first version can be used for tests and maybe
sensitivity studies.