Added Skimmer for mini AODs
Since the miniAOD is not compatable with the normal AOD format I added a
seperate tool chain. As far as I can see all the normal IDs are
fullfilled, except the photon ID conversion. I couldn't find a way to
acess that. But the CutBased ID are stored as flags.