Fix flags, globaltags and filenames to reflect data / mc
This commit will fix the runOnData flags, globaltags and filenames to reflect
either data or mc according to the filename. There has been confusion beforehand,
resulting in the mc file being used for data skimming. To prevent future
misunderstandings, additional documentation as been added to the top of the
files.